Fun stuff!
Added the start of some 'break down effects' with particles... added
into the Mastermind Controller - maybe not the best way to do it-- feel
free to modify. added warp sounds to all the ships. Oh, the Mastermind
Controller says that it has a bunch of unreferenced objects, hence the
inability for the puzzles to currently work. Promise I didn't do it!