Lasers and Asteroids
Now asteroids float through the level, and holding down 'fire' shoots
them. Ideally we tie this function to the main red button, at the
beginning of the game, and after the weapons have been rebooted. They
could be a purely bonus feature, or we could actually make it so the
asteroids could destroy your ship and cause an instant game over, which
would add an extra bit of madness while you're puzzle solving.